Cleanup tests
Removes ComponentResourcePostgresTest and moves its tests to ComponentResourceTest instead. The former pre-dates our choice to focus entirely on Postgres, when most tests still ran against an in-memory H2 DB.
